Added customizations to Errors
