+++ controllers/blogpostController.js
@@ -6,56 +6,7 @@ const Blogpost = require('../models/Blogpost');
 //@route        GET /api/v1/blogposts
 //@access       Public
 exports.getBlogPosts = asyncHandler(async (req, res, next) => {
-  let query;
-
-  // Copy of req.query
-  const reqQuery = { ...req.query };
-
-  // Fields to exclude
-  const removeFields = ['page', 'limit'];
-
-  // Loop over removeFields and delete them from reqQuery
-  removeFields.forEach((param) => delete reqQuery[param]);
-
-  // Searching DB
-  query = Blogpost.find(reqQuery);
-
-  // Pagination
-  const page = parseInt(req.query.page, 10) || 1;
-  const limit = parseInt(req.query.limit, 10) || 10;
-  const startIndex = (page - 1) * limit;
-  const endIndex = page * limit;
-  const total = await Blogpost.countDocuments();
-
-  query = query.skip(startIndex).limit(limit);
-
-  // Execute Query
-  const blogposts = await query;
-
-  // Pagination result
-  const pagination = {};
-
-  if (endIndex < total) {
-    pagination.next = {
-      page: page + 1,
-      // limit,
-    };
-  }
-
-  if (startIndex > 0) {
-    pagination.prev = {
-      page: page - 1,
-      // limit,
-    };
-  }
-
-  res.status(200).json({
-    success: true,
-    count: blogposts.length,
-    pagination,
-    pageLimit: limit,
-    data: blogposts,
-  });
+  res.status(200).json(res.resultCustomizationMiddleware);
 });
 
 // @desc        Create a blogpost
