+++ routes/blogpostRoutes.js
@@ -6,10 +6,15 @@ const {
   updateBlogPost,
   deleteBlogPost,
 } = require('../controllers/blogpostController');
+const Blogpost = require('../models/Blogpost');
+const resultCustomizationMiddleware = require('../middleware/resultCutomizationMiddleware');
 
 const router = express.Router();
 
-router.route('/').get(getBlogPosts).post(createBlogPost);
+router
+  .route('/')
+  .get(resultCustomizationMiddleware(Blogpost), getBlogPosts)
+  .post(createBlogPost);
 
 router
   .route('/:id')
